// frontend/src/api/generated/models/unit-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 荷姿
 * @export
 * @enum {string}
 */

export const UnitType = {
    Case: 'CASE',
    Ball: 'BALL',
    Piece: 'PIECE'
} as const;

export type UnitType = typeof UnitType[keyof typeof UnitType];